+++ eslint.config.mjs
@@ -1,0 +1,8 @@
+import neolutionEslintConfig from "@neolution-ch/eslint-config-neolution";
+
+export default [
+  {
+    ignores: ["**/cypress/"],
+  },
+  ...neolutionEslintConfig.configs.flat["react-library"],
+];